requirements.txt
streamlit
pandas
google-generativeai
pymupdf
git+https://github.com/haven-jeon/PyKoSpacing.git